pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "cachetools>=6.1.0",
    "openai>=1.97.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.3",
    "pytz>=2025.2",
    "requests>=2.32.4",
]
